+++ apps/formio-server/Changelog.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 5.8.7
 ### Added
+ - The ability to change the submission PDF filename that gets attached to the emails.
+
+### Changed
+ - Upgrade formio to 1.34.3 to allow for current endpoints to accept temp tokens.
  - Debugging for LDAP Login action
 
 ## 5.8.6